changed logic of object access
user should have access to object if user is the owner. Therefore, if user == owner then return true to provide access, otherwise return NULL (not FALSE) so as to assure that this method has no impact upon restricted accesses.  If you are not the owner of an object, you may still have access if you are the administrator.
refs TRAC-849